requirements.txt
bs4
openai
pydantic
python-dotenv
requests
typing